% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_close.R
\name{comm_close}
\alias{comm_close}
\title{Comm Close}
\usage{
comm_close(request_msg)
}
\arguments{
\item{request_msg}{A list passed in from \code{doRequest} representing the
deserialized \code{comm_close} message JSON.}
}
\description{
Handler for the comm_close Message Type
}
\examples{
\dontrun{
  request_msg <- list("comm_id"="my_open_comm_id", "data"=list())
  comm_close(request_msg)
}

}
\references{
\url{http://jupyter-client.readthedocs.io/en/latest/messaging.html#tearing-down-comms}
}
\author{
Spencer Aiello
}
