% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_msg.R
\name{comm_msg}
\alias{comm_msg}
\title{Comm Msg}
\usage{
comm_msg(request_msg)
}
\arguments{
\item{request_msg}{A list passed in from \code{doRequest} representing the
deserialized \code{comm_msg} message JSON.}
}
\description{
Handler for the comm_msg Message Type
}
\examples{
\dontrun{
  request_msg <- list("comm_id"="my_open_comm_id", "data"=list())
  comm_request(request_msg)
}

}
\references{
\url{http://jupyter-client.readthedocs.io/en/latest/messaging.html#comm-messages}
}
\author{
Spencer Aiello
}
