% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelspec.R
\name{defaultDisplayName}
\alias{defaultDisplayName}
\title{Create the Default Kernel Display Name}
\usage{
defaultDisplayName()
}
\description{
Create Kernel Display Name
}
\examples{
\dontrun{
  defaultDisplayName()
}

}
\seealso{
\code{\link{installJuniper}}.
}
