% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelspec.R
\name{defaultKernelName}
\alias{defaultKernelName}
\title{Create the Default Kernel Name}
\usage{
defaultKernelName()
}
\description{
Create Kernel Name
}
\examples{
\dontrun{
  defaultKernelName()
}

}
\seealso{
\code{\link{installJuniper}}.
}
