% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_request.R
\name{execute_request}
\alias{execute_request}
\title{Execute Handler}
\usage{
execute_request(request_msg)
}
\arguments{
\item{request_msg}{A list passed in from \code{doRequest} representing the
deserialized \code{execute_request} message JSON.}
}
\value{
A list having names \code{msg_type} and \code{content}. The
  \code{msg_type} is \code{execute_reply}, which corresponds
  to the \code{execute_request} message. The \code{content} field
  complies with the Jupyter wire message protocol specification
  for \code{execute_reply} messages.
}
\description{
Handler for the execute_request Message Type
}
\examples{
\dontrun{
  request_msg <- list("code"="5+5")
  execute_request(request_msg)
}

}
\references{
\url{http://jupyter-client.readthedocs.io/en/latest/messaging.html#execution-results}
}
\author{
Spencer Aiello
}
