% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_request.R
\name{history_request}
\alias{history_request}
\title{History Request Handler}
\usage{
history_request(request_msg)
}
\arguments{
\item{request_msg}{A list passed in from \code{doRequest} representing the
deserialized \code{history_request} message JSON.}
}
\value{
A list having names \code{msg_type} and \code{content}. The
  \code{msg_type} is \code{history_reply}, which corresponds
  to the \code{history_request} message. The \code{content} field
  complies with the Jupyter wire message protocol specification
  for \code{history_reply} messages.
}
\description{
Handler for the history_request Message Type
}
\references{
\url{http://jupyter-client.readthedocs.io/en/latest/messaging.html#history}
}
\author{
Spencer Aiello
}
