% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_complete_request.R
\name{is_complete_request}
\alias{is_complete_request}
\title{Is Complete Handler}
\usage{
is_complete_request(request_msg)
}
\arguments{
\item{request_msg}{A list passed in from \code{doRequest} representing the
deserialized \code{is_complete_request} message JSON.}
}
\value{
A list having names \code{msg_type} and \code{content}. The
  \code{msg_type} is \code{is_complete_reply}, which corresponds
  to the \code{is_complete_request} message. The \code{content} field
  complies with the Jupyter wire message protocol specification
  for \code{is_complete_reply} messages.
}
\description{
Handler for the is_complete_request Message Type
}
\examples{
\dontrun{
  request_msg <- list(code="print(")  'incomplete'
  is_complete_request(request_msg)
}

\dontrun{
  request_msg <- list(code="print(5)")  # 'complete'
  is_complete_request(request_msg)
}

}
\references{
\url{http://jupyter-client.readthedocs.io/en/latest/messaging.html#code-completeness}
}
\author{
Spencer Aiello
}
