% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_device.R
\name{jk_device_defaults}
\alias{jk_device_defaults}
\title{Set JuniperKernel's Graphic Device Settings to Defaults}
\usage{
jk_device_defaults()
}
\value{
Invisibly return a list of the settings.
}
\description{
Reset JuniperKernel's Graphic Settings
}
\details{
Use this method to set the graphics settings used by JuniperKernel.
}
\author{
Spencer Aiello
}
