% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_info_request.R
\name{kernel_info_request}
\alias{kernel_info_request}
\title{Kernel Info Request Handler}
\usage{
kernel_info_request(request_msg)
}
\arguments{
\item{request_msg}{A list passed in from \code{doRequest} representing the
deserialized \code{kernel_info_request} message JSON.}
}
\value{
A list having names \code{msg_type} and \code{content}. The
  \code{msg_type} is \code{kernel_info_reply}, which corresponds
  to the \code{kernel_info_request} message. The \code{content} field
  complies with the Jupyter wire message protocol specification
  for \code{kernel_info_reply} messages.
}
\description{
Response to the \code{kernel_info_request} Message Type
}
\examples{
\dontrun{
  kernel_info_request(list())
}

}
\references{
\url{http://jupyter-client.readthedocs.io/en/latest/messaging.html#kernel-info}
}
\author{
Spencer Aiello
}
