% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelspec.R
\name{listKernels}
\alias{listKernels}
\title{List Installed Jupyter Kernels}
\usage{
listKernels()
}
\description{
List Kernels
}
\details{
Prints the currently installed kernels and their install locations.
}
\examples{
\dontrun{
  listKernels()
}

}
