% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_open.R
\name{comm_open}
\alias{comm_open}
\title{Comm Open}
\usage{
comm_open(request_msg)
}
\arguments{
\item{request_msg}{A list passed in from \code{doRequest} representing the
deserialized \code{comm_open} message JSON.}
}
\description{
Handler for the comm_open Message Type
}
\examples{
\dontrun{
  request_msg <- list("comm_id"="uniq_comm_id", "target_name"="my_comm", "data"=list())
  comm_open(request_msg)
}

}
\references{
\url{http://jupyter-client.readthedocs.io/en/latest/messaging.html#opening-a-comm}
}
\author{
Spencer Aiello
}
